update installer to create global link

+++ installer.sh
@@ -55,10 +55,32 @@ echo "Installing dependencies..."
 
 check_app_or_install "MakeMKV" "makemkv"
 check_app_or_install "Wine Stable" "wine-stable"
-ln -s /Applications/Wine\ Stable.app/Contents/Resources/wine/bin/wine "$BREW_PATH/wine"
+if [ ! -L "$BREW_PATH/wine" ]; then
+    ln -s /Applications/Wine\ Stable.app/Contents/Resources/wine/bin/wine "$BREW_PATH/wine"
+fi
+
+echo "Creating a virtual environment for BD_to_AVP..."
+VENV_PATH="$HOME/.bd_to_avp_venv"
+python3.12 -m venv "$VENV_PATH" || handle_error "Failed to create a virtual environment"
+
+echo "Activating the virtual environment..."
+source "$VENV_PATH/bin/activate"
 
 echo "Installing BD_to_AVP from PyPI..."
-python3 -m pip install bd_to_avp || handle_error "Failed to install BD_to_AVP from PyPI"
+pip install bd_to_avp || handle_error "Failed to install BD_to_AVP from PyPI"
+
+echo "Making BD_to_AVP executable accessible system-wide..."
+
+EXECUTABLE_PATH="$VENV_PATH/bin/bd-to-avp"
+SYSTEM_WIDE_PATH="/opt/homebrew/bin/bd-to-avp"
+
+if [ -L "$SYSTEM_WIDE_PATH" ] || [ -e "$SYSTEM_WIDE_PATH" ]; then
+    echo "Existing BD_to_AVP executable or symlink found. Removing..."
+    rm -f "$SYSTEM_WIDE_PATH"
+fi
+
+ln -s "$EXECUTABLE_PATH" "$SYSTEM_WIDE_PATH" || handle_error "Failed to link BD_to_AVP executable system-wide"
+echo "BD_to_AVP is now accessible system-wide as 'bd-to-avp'"
 
 echo "Installing Rosetta 2 (if required)..."
 if arch -x86_64 true 2>/dev/null; then
@@ -69,8 +91,6 @@ else
 fi
 
 $BREW_PATH/wineboot &> /dev/null;
-
-
 
 echo "BD_to_AVP environment setup complete. Refresh your terminal's environment to use new paths."
 echo "1. Refresh environment: source $HOME/.zshrc"
